# backend/openapi.yml
...
openapi: 3.1.0
info:
  title: FocusMe Cloud API
  version: 1.0.0
  description: |
    Cloud backend API for FocusMe — provides user authentication,
    cross-device plan synchronization, and family dashboard features.

    ## Authentication

    Most endpoints require a Bearer JWT access token in the
    `Authorization` header. Obtain tokens via `/auth/login` or
    `/auth/register`. Access tokens expire after 15 minutes;
    use `/auth/refresh` to rotate.

    ## Rate Limiting

    - Auth routes: **10 requests/minute** per IP
    - All other routes: **100 requests/minute** per IP
    - Exceeded → `429 Too Many Requests` with `Retry-After` header
  contact:
    name: FocusMe Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.focusme.app/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development

# ── Security Schemes ──────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from /auth/login or /auth/register.
        Include as `Authorization: Bearer <token>`.

  # ── Reusable Schemas ──────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code constant for programmatic handling
          example: "AUTH_INVALID_CREDENTIALS"
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EMAIL_EXISTS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_REFRESH_REUSE
            - SYNC_VERSION_CONFLICT
            - SYNC_PLAN_NOT_FOUND
            - FAMILY_INVITE_EXPIRED
            - FAMILY_ALREADY_MEMBER
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message (do not match on this — use code instead)
          example: "Authentication failed: invalid credentials"
        details:
          type: object
          description: Optional structured error context (e.g., field-level validation errors)

    # ── Health ──────────────────────────────────────────────
...
    Plan:
      type: object
      required: [id, user_id, local_id, plan_json, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        local_id:
          type: string
        plan_json:
          type: object
        version:
          type: integer
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Sync ────────────────────────────────────────────────
...
    CreateFamilyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Family group display name
          example: "Smith Family"
    FamilyGroup:
      type: object
      required: [id, name, owner_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    FamilyMember:
      type: object
      required: [id, user_id, role, status]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum: [owner, member]
        status:
          type: string
          enum: [pending, accepted]
    InviteRequest:
      type: object
      required: [email, group_id]
      properties:
        email:
          type: string
          format: email
          description: Email of the user to invite
          example: "family@example.com"
        group_id:
          type: string
          format: uuid
    AcceptInviteRequest:
      type: object
      required: [invite_token]
      properties:
        invite_token:
          type: string
          format: uuid
          description: Token from the invite
    InviteResponse:
      type: object
      required: [invite_token, expires_at]
      properties:
        invite_token:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
    SharePlanRequest:
      type: object
      required: [group_id]
      properties:
        group_id:
          type: string
          format: uuid
    DashboardResponse:
      type: object
      required: [group, members, shared_plans, member_count]
      properties:
        group:
          $ref: "#/components/schemas/FamilyGroup"
        members:
          type: array
          items:
            $ref: "#/components/schemas/FamilyMember"
        shared_plans:
          type: array
          items:
            $ref: "#/components/schemas/Plan"
        member_count:
          type: integer
          example: 4

# ── Paths ──────────────────────────────────────────────────────
paths:
...
  /family:
    post:
      operationId: createFamily
      summary: Create a new family group
      description: |
        Creates a family group with the authenticated user as owner.
        The owner is automatically added as a member with `accepted` status.
      tags: [Family]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyRequest"
      responses:
        "201":
          description: Family group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyGroup"
        "400":
          description: Validation error (empty name)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Validation error: family name cannot be empty"
                code: 400
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer

    get:
      operationId: listFamilies
      summary: List user's family groups
      description: Returns all family groups the authenticated user belongs to.
      tags: [Family]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Family groups retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyGroup"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
  /family/members:
    get:
      operationId: listFamilyMembers
      summary: List all family members across groups
      description: Returns members of all family groups the user belongs to.
      tags: [Family]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Members retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
  /family/invite:
    post:
      operationId: inviteFamilyMember
      summary: Invite a user to a family group
      description: |
        Creates a pending invitation with a UUID token that expires
        after 7 days. Only the group owner can send invites.
      tags: [Family]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRequest"
      responses:
        "200":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only the group owner can invite members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden: only the group owner can invite members"
                code: 403
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
  /family/invite/accept:
    post:
      operationId: acceptInvite
      summary: Accept a family group invite
      description: |
        Validates the invite token (must not be expired), marks the
        membership as `accepted`, and clears the invite token.
      tags: [Family]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
      responses:
        "200":
          description: Invite accepted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invite accepted"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invalid or expired invite token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
  /family/plans/share/{plan_id}:
    post:
      operationId: sharePlan
      summary: Share a plan with a family group
      description: |
        Creates a shared copy of the specified plan for the given family group.
        The plan is copied, not referenced, to avoid cross-user conflicts.
      tags: [Family]
      security:
        - BearerAuth: []
      parameters:
        - name: plan_id
          in: path
          required: true
          description: UUID of the plan to share
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharePlanRequest"
      responses:
        "200":
          description: Plan shared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Plan shared with family group"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Plan or family group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
  /family/dashboard:
    get:
      operationId: familyDashboard
      summary: Family activity summary dashboard
      description: |
        Returns an aggregate view of the user's primary family group,
        including all members, shared plans, and member count.
      tags: [Family]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard data retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User is not a member of any family group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not found: user is not a member of any family group"
                code: 404
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer

# ── Tags ──────────────────────────────────────────────────────
...